+++ src/Catrobat/Services/ScreenshotRepository.php
@@ -3,6 +3,7 @@
 namespace App\Catrobat\Services;
 
 use App\Catrobat\Exceptions\InvalidStorageDirectoryException;
+use App\Utils\Utils;
 use Imagick;
 use ImagickException;
 use Symfony\Component\Filesystem\Filesystem;
@@ -118,9 +119,10 @@ class ScreenshotRepository
 
   public function getScreenshotWebPath(string $id): string
   {
-    if (file_exists($this->screenshot_dir.$this->generateFileNameFromId($id)))
+    $filename = $this->screenshot_dir.$this->generateFileNameFromId($id);
+    if (file_exists($filename))
     {
-      return $this->screenshot_path.$this->generateFileNameFromId($id);
+      return $this->screenshot_path.$this->generateFileNameFromId($id).Utils::getTimestampParameter($filename);
     }
 
     return self::DEFAULT_SCREENSHOT;
@@ -131,9 +133,10 @@ class ScreenshotRepository
    */
   public function getThumbnailWebPath($id): string
   {
-    if (file_exists($this->thumbnail_dir.$this->generateFileNameFromId((string) $id)))
+    $filename = $this->thumbnail_dir.$this->generateFileNameFromId((string) $id);
+    if (file_exists($filename))
     {
-      return $this->thumbnail_path.$this->generateFileNameFromId((string) $id);
+      return $this->thumbnail_path.$this->generateFileNameFromId((string) $id).Utils::getTimestampParameter($filename);
     }
 
     return self::DEFAULT_THUMBNAIL;
@@ -243,7 +246,7 @@ class ScreenshotRepository
    */
   public function deleteTempFiles(): void
   {
-    $this->removeDirectory($this->tmp_dir);
+    Utils::removeDirectory($this->tmp_dir);
   }
 
   /**
@@ -299,36 +302,5 @@ class ScreenshotRepository
     $thumb->writeImage($filename);
     chmod($filename, 0777);
     $thumb->destroy();
-  }
-
-  private function removeDirectory(string $directory): void
-  {
-    foreach (glob(sprintf('%s*', $directory)) as $file)
-    {
-      if (is_dir($file))
-      {
-        $this->recursiveRemoveDirectory($file);
-      }
-      else
-      {
-        unlink($file);
-      }
-    }
-  }
-
-  private function recursiveRemoveDirectory(string $directory): void
-  {
-    foreach (glob(sprintf('%s/*', $directory)) as $file)
-    {
-      if (is_dir($file))
-      {
-        $this->recursiveRemoveDirectory($file);
-      }
-      else
-      {
-        unlink($file);
-      }
-    }
-    rmdir($directory);
   }
 }