+++ tests/phpUnit/Hook/RefreshTestEnvHook.php
@@ -2,11 +2,15 @@
 
 namespace Tests\phpUnit\Hook;
 
-use App\Commands\Helpers\CommandHelper;
+use App\Catrobat\Services\TestEnv\DataFixtures\ProjectDataFixtures;
+use App\Catrobat\Services\TestEnv\DataFixtures\UserDataFixtures;
+use App\Kernel;
+use Doctrine\ORM\EntityManagerInterface;
 use PHPUnit\Runner\BeforeFirstTestHook;
 use PHPUnit\Runner\BeforeTestHook;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Finder\Finder;
+use Tests\behat\context\RefreshEnvironmentContext;
 
 class RefreshTestEnvHook implements BeforeTestHook, BeforeFirstTestHook
 {
@@ -25,14 +29,33 @@ class RefreshTestEnvHook implements BeforeTestHook, BeforeFirstTestHook
 
   public function executeBeforeFirstTest(): void
   {
-    CommandHelper::executeShellCommand(
-      ['bin/console', 'catrobat:test:generate', '--env=test', '--no-interaction'], [], 'Generating test data'
-    );
+    RefreshEnvironmentContext::prepare();
   }
 
   public function executeBeforeTest(string $test): void
   {
     $this->emptyDirectory(self::$CACHE_DIR);
+  }
+
+  // Unit tests for now rarely need a database rollback. For performance reasons we only execute it when needed.
+  public static function databaseRollback(): void
+  {
+    $kernel = new Kernel($_SERVER['APP_ENV'], (bool) $_SERVER['APP_DEBUG']);
+    $kernel->boot();
+
+    /* @var EntityManagerInterface $em */
+    $em = $kernel->getContainer()->get('doctrine')->getManager();
+
+    $em->getConnection()->query('SET FOREIGN_KEY_CHECKS=0');
+    foreach ($em->getConnection()->getSchemaManager()->listTableNames() as $tableName)
+    {
+      $q = $em->getConnection()->getDatabasePlatform()->getTruncateTableSql($tableName);
+      $em->getConnection()->executeUpdate($q);
+    }
+    $em->getConnection()->query('SET FOREIGN_KEY_CHECKS=1');
+
+    ProjectDataFixtures::clear();
+    UserDataFixtures::clear();
   }
 
   private function emptyDirectory(string $directory): void